pull get_jsonc2 out of fitness
