+++ worker_health/worker_health/utils.py
@@ -3,6 +3,8 @@ import logging
 import pprint
 import subprocess
 from urllib.request import urlopen
+from requests.adapters import HTTPAdapter
+from urllib3.util import Retry
 
 import requests
 
@@ -122,3 +124,55 @@ def fetch_url(url):
         return url, response.read(), None
     except Exception as e:
         return url, None, e
+
+
+# https://www.peterbe.com/plog/best-practice-with-retries-with-requests
+def requests_retry_session(
+    retries=3,
+    backoff_factor=0.3,
+    status_forcelist=(500, 502, 504),
+    session=None,
+):
+    session = session or requests.Session()
+    retry = Retry(
+        total=retries,
+        read=retries,
+        connect=retries,
+        backoff_factor=backoff_factor,
+        status_forcelist=status_forcelist,
+    )
+    adapter = HTTPAdapter(max_retries=retry)
+    session.mount("http://", adapter)
+    session.mount("https://", adapter)
+    return session
+
+
+# returns (url, response, exception)
+def get_jsonc2(an_url, verbosity=0):
+    headers = {"User-Agent": USER_AGENT_STRING}
+    retries_left = 2
+
+    while retries_left >= 0:
+        if verbosity > 2:
+            print(an_url)
+        response = requests_retry_session().get(an_url, headers=headers)
+        result = response.text
+        try:
+            output = json.loads(result)
+            # will only break on good decode
+            break
+        except json.decoder.JSONDecodeError as e:
+            logger.warning("json decode error. input: %s" % result)
+            if retries_left == 0:
+                return an_url, None, e
+        print("request failure, manual retry")
+        retries_left -= 1
+
+    while "continuationToken" in output:
+        payload = {"continuationToken": output["continuationToken"]}
+        if verbosity > 2:
+            print("%s, %s" % (an_url, output["continuationToken"]))
+        response = requests_retry_session().get(an_url, headers=headers, params=payload)
+        result = response.text
+        output = json.loads(result)
+    return an_url, output, None
